+++ CHANGELOG.md
@@ -11,6 +11,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.2.1] - 2021-04-06
 
+## Added
+
++ added distances and directions functions
+
+## Changed
+
++ rename directions to units
+
 ## [0.2.0] - 2021-04-06
 
 ## Added